pyproject.toml
[project]
name = "moltbot-helpers"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "numpy>=1.24",
    "thefuzz[speedup]>=0.20",
    "psutil>=5.9.0",
    "watchdog>=3.0.0",
    "psycopg2-binary>=2.9.0",
    "markdown>=3.4.0",
    "beautifulsoup4>=4.12.0"
]

[project.scripts]
fp = "fp.cli:main"
fp_pg = "fp_pg.cli:main"
seek = "seek.cli:main"
seek_pg = "seek_pg.cli:main"
seek_pg_old = "seek.cli_pg:main"
pt = "pt.cli:main"
pt_pg = "pt_pg.cli:main"
gm = "gmail.cli:main"
md2pg = "markdown_processor.cli:main"
lookup = "lookup.cli:main"
integrate = "integration.cli:main"
gc = "gcal.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/fp", "src/seek", "src/seek_pg", "src/pt", "src/pt_pg", "src/lookup", "src/integration", "src/gcal", "src/gmail", "src/markdown_processor", "src/backup", "src/sync", "src/process_manager"]

[tool.hatch.build.force-include]